+++ src/main/java/com/github/whileloop/rest4j/HttpResponse.java
@@ -3,7 +3,6 @@ package com.github.whileloop.rest4j;
 import java.io.IOException;
 import java.io.OutputStream;
 
-import static com.github.whileloop.rest4j.HttpRequest.GSON;
 
 /**
  * Created by aalves on 12/18/17
@@ -12,7 +11,7 @@ public interface HttpResponse extends HttpContext {
     OutputStream getRawBody();
 
     default void write(Object object) throws IOException {
-        write(GSON.toJson(object));
+        write(JsonUtils.toJson(object));
     }
 
     void write(byte[] content) throws IOException;