+++ CHANGELOG.md
@@ -6,8 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.2] yyyy-mm-dd
+## [1.0.3] yyyy-mm-dd
 
+## [1.0.2] 2018-07-25
+### Added
+- More Json helper methods in [`HttpRequest`](src/main/java/com/github/whileloop/rest4j/HttpRequest.java)
+- More Json helper methods in [`HttpResponse`](src/main/java/com/github/whileloop/rest4j/HttpResponse.java)
+ 
 ## [1.0.1] 2018-07-24
 ### Added
 - Json helper methods in [`HttpRequest`](src/main/java/com/github/whileloop/rest4j/HttpRequest.java)
@@ -24,7 +29,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [1.0.0]: https://github.com/while-loop/rest4j/releases/tag/v1.0.0
 [1.0.1]: https://github.com/while-loop/rest4j/compare/v1.0.0...v1.0.1
-[1.0.2]: https://github.com/while-loop/rest4j/compare/v1.0.1...master
-[unreleased]: https://github.com/while-loop/rest4j/compare/v1.0.1...master
+[1.0.2]: https://github.com/while-loop/rest4j/compare/v1.0.1...v1.0.2
+[1.0.3]: https://github.com/while-loop/rest4j/compare/v1.0.2...develop
+[unreleased]: https://github.com/while-loop/rest4j/compare/v1.0.2...develop
 
 [comment]: # (Added, Changed, Removed)